rename rule and add docs
