+++ packages/eslint-plugin-query/src/configs/index.test.ts
@@ -10,7 +10,7 @@ describe('configs', () => {
           ],
           "rules": {
             "@tanstack/query/exhaustive-deps": "error",
-            "@tanstack/query/no-callbacks": "error",
+            "@tanstack/query/no-deprecated-options": "error",
             "@tanstack/query/prefer-query-object-syntax": "error",
             "@tanstack/query/stable-query-client": "error",
           },
