+++ packages/eslint-plugin-query/src/rules/no-deprecated-options/no-deprecated-options.test.ts
@@ -1,6 +1,6 @@
 import { ESLintUtils } from '@typescript-eslint/utils'
 import { normalizeIndent } from '../../utils/test-utils'
-import { name, rule } from './no-callbacks.rule'
+import { name, rule } from './no-deprecated-options.rule'
 
 const ruleTester = new ESLintUtils.RuleTester({
   parser: '@typescript-eslint/parser',
